Start Menu_(Read Description)
The start menu includes
-New Game
-Options
-Exit

New Game :
It will load the samplescene on clicking.

Options :
-Resolution_Dropdown
-Quality_Setting_Dropdown
-Fullscreen
-Volume

Resolution_Dropdown :
It checks the available resolutions on the platform and provides the options for the player so that the optimal resolution can be selected by the user.

Quality_Settings_Dropdown
It provides the user with quality drop down
-ultra
-veryhigh
-high
-medium
-low
-verylow

FullScreen :
Toggles teh full screen for the game

Volume :
It is a slider which sets the volume of the game

Exit :
Onclick exits the game.